requirements.txt
# QLib
pyqlib

# 数据处理
pandas>=1.3.0
numpy>=1.20.0

# 机器学习
scikit-learn>=0.24.0

# 数据源
tushare
baostock

# 工具
pyyaml
tqdm

# 日志
loguru
